fix(asmdef): change asmdef to Unity 2018.3 format
The asmdef file was created in a newer version of Unity where the
format is different, which will cause errors if this repo is used
in an earlier version of Unity where the new format is not supported.

This format will be auto updated to the new format by Unity.